♻️ refactor: refactor page name
