+++ src/app/chat/mobile/page.page.tsx
@@ -1,4 +1,4 @@
-import Page from '@/pages/market/index.page';
+import Page from '@/pages/chat/mobile';
 
 const Index = () => {
   return <Page />;
