+++ src/app/chat/page.page.tsx
@@ -1,4 +1,4 @@
-import Page from '@/pages/chat/mobile/index.page';
+import Page from '@/pages/chat';
 
 const Index = () => {
   return <Page />;
