+++ src/app/chat/settings/page.page.tsx
@@ -1,4 +1,4 @@
-import Page from '@/pages/chat/index.page';
+import Page from '@/pages/chat/setting';
 
 const Index = () => {
   return <Page />;
